Refactor to base filter params dto - models - mapper JC-655
