+++ libs/core/dtos/base-filter-params.dto.ts
@@ -1,0 +1,23 @@
+/** Base filter params DTO. */
+export namespace BaseFilterParamsDto {
+
+	/** Search. */
+	export type Search = {
+
+		/** Search filter. */
+		readonly search?: string;
+	};
+
+	/** Pagination. */
+	export type Pagination = {
+
+		/** Offset. */
+		readonly offset?: number;
+
+		/** Page size. */
+		readonly limit?: number;
+	};
+
+	/** Combined params. */
+	export type Combined = Search & Pagination;
+}
